creation of a unique inference script
